requirements.txt
antlr4-python3-runtime==4.8
Django==2.2.6
JPype1==0.7.0
ontospy==1.9.8.3
wordninja==2.0.0